Avoid delivery of history to volatile reader
When a remote writer is discovery, a proxy_writer object representation
that writer is created without yet having any knowledge of what the
current sequence number for that writer is.  If a local reader is
matched with that proxy writer before a Heartbeat has been recevied and
this sequence number information is known, all historical data will be
made available to that reader, even if it is volatile.

By treating the first Heartbeat specially, by moving the next sequence
number to be delivered as fresh data forward to the next sequence
number, retrieval of historical data is avoided.  Transient-local
readers have a separate ("out-of-sync") route to request it anyway.

Signed-off-by: Erik Boasson <eb@ilities.com>