Fix conversion of {sec,nsec} to msec in timedwait on Windows
Internally time stamps and durations are all in nanoseconds, but the
platform abstraction uses {sec,nsec} (essentially a struct timespec) and
Windows uses milliseconds.  The conversion to milliseconds with upwards
rounding was broken, adding ~1s to each timeout.  In most of the handful
of uses the effect is minor in practice, but it does matter a lot in the
scheduling of Heartbeat and AckNack messages, e.g., by causing a simple
throughput test to exhibit periodic drops in throughput.

Signed-off-by: Erik Boasson <eb@ilities.com>