MPT driver incorrectly reporting a waitpid error
The multi-process test driver program waits until all its children have
stopped (or timeout) by calling ddsrt_proc_waitpids in a loop.  The way
the loop is written, it can try once more when all children have already
been waited for, in which case it reports a spurious error.

This fixes that problem by checking whether the error code was to be
expected.

Signed-off-by: Erik Boasson <eb@ilities.com>